requirements.txt
fastapi
uvicorn
langchain>=0.3
langchain-aws>=0.2
langchain-core>=0.3
langgraph>=0.2
psycopg2-binary
python-dotenv
boto3
matplotlib
pandas
